+++ app/decorators/residency_decorator.rb
@@ -11,7 +11,7 @@ class ResidencyDecorator < Draper::Decorator
     status_tag = h.link_to(object.property, class: 'no-link-icon' ) do
       resident_status_tag(tooltip: tooltip)
     end
-    if object.verified_on?
+    if object.verified?
       status_tag
       # h.concat(h.content_tag(:span, '✓', title: 'Verified at this address'))
     else
@@ -30,26 +30,8 @@ class ResidencyDecorator < Draper::Decorator
     h.link_to(object.resident, resident.to_s)
   end
 
-  def resident_status_character
-    return '' if is_minor? # child
-
-    case resident_status
-    when nil
-      '⁇'
-    when :border.to_s
-      '' # people-robbery
-    when /owner/, :trustee.to_s
-      '' # gavel
-    when :renter.to_s
-      '' # suitcase
-    when :dependent.to_s
-      # '' # family, pro
-      '' # user-graduate
-    when :significant_other.to_s
-      '' # user-group
-    else
-      raise NotImplementedError, "Unknown resident_status: #{resident_status.inspect}"
-    end
+  def resident_status_character(resident_status = object.resident_status)
+    h.resident_status_character(resident_status, is_minor?)
   end
 
   def resident_status_i18n
@@ -61,36 +43,46 @@ class ResidencyDecorator < Draper::Decorator
     end
   end
 
-  # def resident_status_link_tag(tooltip: self.tooltip(show_property: false, show_resident: false))
-  #   h.link_to(object.resident, class: 'fas no-link-icon') do
-  #     resident_status_tag(tooltip: tooltip)
-  #   end
-  # end
-  def resident_status_link_tag(tooltip: self.tooltip(show_property: false))
-    # h.link_to(object.resident, resident.to_s)
-    status_tag = h.link_to(object.resident, class: 'no-link-icon' ) do
+  def resident_status_link_tag(show_property: false, show_resident: false)
+    tooltip = self.tooltip(show_property: show_property, show_resident: show_resident)
+    h.link_to(object.resident, class: 'fas no-link-icon') do
       resident_status_tag(tooltip: tooltip)
-    end
-    if object.verified_on?
-      status_tag
-      # h.concat(h.content_tag(:span, '✓', title: 'Verified at this address'))
-    else
-      h.content_tag(:div, class: 'group') do
-        h.concat(status_tag)
-        h.concat(helpers.simple_form_for(object, remote: true) do |f|
-          f.input :verified_on, as: :hidden, input_html: { value: DateTime.now }
-        end)
-        tooltip = "Click to verify: #{object.resident.full_name.inspect} at #{object.property.street_address.inspect}"
-        h.concat(helpers.check_box_tag("#{object.to_global_id}_verified", object.verified?, object.verified?, onclick: 'updateVerifiedOn(this)', data: {tooltip: tooltip} ))
-      end
     end
   end
 
-  def resident_status_tag(tooltip: resident_status_i18n)
+  def resident_updatable_status_link_tag(tooltip: self.tooltip(show_property: false))
+    if object.resident_status
+      return h.link_to(object.resident, class: 'no-link-icon' ) do
+        resident_status_tag(resident_status: object.resident_status, tooltip: tooltip)
+      end
+    end
+
+    # present all, with ability to assign one
+    list_item_tags = Residency.resident_statuses.collect do |status_key, status_name|
+      h.content_tag(:li, class: 'group') do
+        h.concat(h.content_tag(:span, class: 'fas icon', data: {tooltip: status_name}) do
+          h.resident_status_character(status_key, false)
+        end)
+        h.concat(status_name)
+        h.concat(helpers.simple_form_for(object, remote: true) do |f|
+          h.concat(f.input :resident_status, as: :hidden, input_html: { value: status_name })
+          h.concat(f.input :verified_on, as: :hidden, input_html: { value: Time.now })
+        end)
+        tooltip = "Click to assign: #{object.resident.full_name.inspect} as #{status_name} at #{object.property.street_address.inspect}"
+        # h.concat(helpers.check_box_tag("#{object.to_global_id}_verified", object.verified?, object.verified?, onclick: 'updateVerifiedOn(this)', data: {tooltip: tooltip} ))
+        h.concat(helpers.check_box_tag("#{object.to_global_id}_resident_status", status_name, false, onclick: 'updateResidentStatus(this)', data: {tooltip: tooltip} ))
+      end
+    end
+    h.content_tag(:ul, class: 'horizontal-list') do
+      h.safe_join(list_item_tags)
+    end
+  end
+
+  def resident_status_tag(resident_status: object.resident_status, tooltip: resident_status_i18n)
     css_classes = %w[fas icon]
     css_classes << 'primary_residence' if primary_residence?
     css_classes << 'second_home' unless primary_residence?
-    h.content_tag(:span, resident_status_character, class: css_classes.join(' '), data: {tooltip: tooltip})
+    h.content_tag(:span, resident_status_character(resident_status), class: css_classes.join(' '), data: {tooltip: tooltip})
   end
 
   def resident_status_icon
@@ -115,7 +107,7 @@ class ResidencyDecorator < Draper::Decorator
     info << resident.to_s if show_resident
     info << '2nd Home' unless primary_residence?
     info << resident_status_i18n
-    info << 'verified' if verified_on?
+    info << 'verified' if verified?
     info.compact.join(", ")
   end
 end